PLATUI-2190: update README and remove silencer plugin
Describe what happens when a user revokes previously granted tracking
consent. We got a question asking if the cookies needed to be removed
and Holly confirmed with Ori that they don't need to be removed to
comply with PECR rules.

Additionally removed the silencer plugin because it wasn't available for
 the latest version of scala that pr-commenter suggested. Then went
 through and fixed up some warnings we were getting.